[cleanup] renames functions to clarify code
