+++ client/reducers/PlayersStatusesReducer.js
