+++ client/reducers/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 import SingleGameReducer from './SingleGameReducer';
 import MultiGameStateReducer from './MultiGameStateReducer';
 import MultiTimerReducer from './MultiTimerReducer';
-import MultiGameProgressReducer from './MultiGameProgressReducer';
+import PlayersStatusesReducer from './PlayersStatusesReducer';
 import CountingDownReducer from './CountingDownReducer';
 import GameTimeReducer from './GameTimeReducer';
 
@@ -10,7 +10,7 @@ const rootReducer = combineReducers({
   singleGame: SingleGameReducer,
   multiGameState: MultiGameStateReducer,
   multiTimer: MultiTimerReducer,
-  multiGameProgress: MultiGameProgressReducer,
+  playersStatuses: PlayersStatusesReducer,
   countingDown: CountingDownReducer,
   gameTime: GameTimeReducer
 });
